refactor: import the 2FA cancellation guard from its leaf module (#7609)
* refactor: import the 2FA cancellation guard from its leaf module

* refactor: drop the 2FA cancellation re-export from the twoFactor barrel

* test: drop the dead requireActual spread from the twoFactor mock

* refactor: move the 2FA service into its own folder and TWO_FACTOR into constants
