+++ app/containers/TwoFactor/index.test.tsx
@@ -1,7 +1,8 @@
 import { act, fireEvent, render, waitFor } from '@testing-library/react-native';
 
 import TwoFactor from '.';
-import { isTwoFactorCancelled, twoFactor } from '../../lib/services/twoFactor';
+import { twoFactor } from '../../lib/services/twoFactor/twoFactor';
+import { isTwoFactorCancelled } from '../../lib/services/twoFactor/twoFactorCancelled';
 
 jest.mock('../../lib/services/restApi', () => ({
 	sendEmailCode: jest.fn()
