+++ app/lib/constants/twoFactor.ts
@@ -1,0 +1,1 @@
+export const TWO_FACTOR = 'TWO_FACTOR';
