+++ app/views/ChangePasswordView/index.test.tsx
@@ -4,8 +4,8 @@ import { useDispatch } from 'react-redux';
 import ChangePasswordView from './index';
 import { useAppSelector } from '../../lib/hooks/useAppSelector';
 import { saveUserProfile } from '../../lib/services/restApi';
-import { twoFactor } from '../../lib/services/twoFactor';
-import { TwoFactorCancelledError } from '../../lib/services/twoFactor';
+import { twoFactor } from '../../lib/services/twoFactor/twoFactor';
+import { TwoFactorCancelledError } from '../../lib/services/twoFactor/twoFactorCancelled';
 import handleSaveUserProfileError from '../../lib/methods/helpers/handleSaveUserProfileError';
 import { TwoFactorMethods } from '../../definitions/ITotp';
 
@@ -22,8 +22,7 @@ jest.mock('../../lib/services/restApi', () => ({
 	setPassword: jest.fn()
 }));
 
-jest.mock('../../lib/services/twoFactor', () => ({
-	...jest.requireActual('../../lib/services/twoFactor'),
+jest.mock('../../lib/services/twoFactor/twoFactor', () => ({
 	twoFactor: jest.fn()
 }));
