+++ app/views/ProfileView/methods/logoutOtherLocations.ts
@@ -4,7 +4,7 @@ import EventEmitter from '../../../lib/methods/helpers/events';
 import { showConfirmationAlert } from '../../../lib/methods/helpers';
 import { events, logEvent } from '../../../lib/methods/helpers/log';
 import { logoutOtherLocations as logoutOtherLocationsService } from '../../../lib/services/restApi';
-import { isTwoFactorCancelled } from '../../../lib/services/twoFactor';
+import { isTwoFactorCancelled } from '../../../lib/services/twoFactor/twoFactorCancelled';
 
 const logoutOtherLocations = () => {
 	logEvent(events.PL_OTHER_LOCATIONS);
