+++ app/views/SetUsernameView.tsx
@@ -20,7 +20,7 @@ import { showErrorAlert } from '../lib/methods/helpers';
 import scrollPersistTaps from '../lib/methods/helpers/scrollPersistTaps';
 import sharedStyles from './Styles';
 import { getUsernameSuggestion, saveUserProfile } from '../lib/services/restApi';
-import { isTwoFactorCancelled } from '../lib/services/twoFactor';
+import { isTwoFactorCancelled } from '../lib/services/twoFactor/twoFactorCancelled';
 import { useAppSelector } from '../lib/hooks/useAppSelector';
 
 const styles = StyleSheet.create({
